requirements.txt
#requirements.txt

streamlit
faiss-cpu
numpy
sentence-transformers
torch
transformers
scikit-learn
requests

